+++ packages/dart/CHANGELOG.md
@@ -2,6 +2,6 @@
 
 All notable changes to this package will be documented in this file.
 
-## 1.10.3
+## 1.10.4
 
 - Initial release.